chore: front-end routes are screens
