+++ src/components/App/index.tsx
@@ -1,5 +1,5 @@
 import isNil from "lodash/isNil";
-import * as React from "react";
+import React from "react";
 import { Outlet, useOutletContext } from "react-router-dom";
 import {
   io,
