+++ src/components/GameSettings/index.tsx
@@ -2,7 +2,7 @@ import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import ToggleButton from "@mui/material/ToggleButton";
 import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
-import * as React from 'react';
+import React from 'react';
 
 import { useSocket } from "../App";
