+++ src/helpers/index.ts
@@ -1,0 +1,9 @@
+export function generateAlphaString(stringLength: number) {
+  let result = "";
+  const characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+  const charactersLength = characters.length;
+  for (let i = 0; i < stringLength; i++) {
+    result += characters.charAt(Math.floor(Math.random() * charactersLength));
+  }
+  return result;
+}
