+++ src/screens/Game/styles.ts
@@ -1,0 +1,13 @@
+import React from "react";
+
+export const appBody: React.CSSProperties = {
+  alignItems: "center",
+  height: "calc(100vh - 60px)",
+  display: "flex",
+  flexDirection: "column",
+  fontFamily: "'Esteban', serif",
+  fontSize: "18px",
+  flexWrap: "nowrap",
+  justifyContent: "flex-start",
+  width: "99vw",
+};
