+++ src/screens/Lobby/index.tsx
@@ -1,0 +1,15 @@
+import GameSettings from "../../components/GameSettings";
+import GameTransition from "../../components/GameTransition";
+import UserTable from "../../components/UserTable";
+
+function Lobby() {
+  return (
+    <div >
+      <UserTable />
+      <GameSettings />
+      <GameTransition />
+    </div>
+  );
+}
+
+export default Lobby;
